requirements.txt
streamlit
supabase
python-dotenv
pandas
requests
bcrypt
pydantic
ruff
mypy
pytest
